+++ packages/eightballer/skills/qs_orchestrator_abci/handlers.py
@@ -17,7 +17,7 @@
 #
 # ------------------------------------------------------------------------------
 
-"""This module contains the handlers for the skill of OrchestratorAbciApp."""
+"""This module contains the handlers for the skill of SQOrchestratorAbciApp."""
 
 from packages.valory.skills.abstract_round_abci.handlers import (
     ABCIRoundHandler as BaseABCIRoundHandler,
