+++ packages/eightballer/skills/qs_orchestrator_abci/skill.yaml
@@ -1,4 +1,4 @@
-name: orchestrator_abci
+name: qs_orchestrator_abci
 author: eightballer
 version: 0.1.0
 type: skill
@@ -6,31 +6,31 @@ description: The scaffold skill is a scaffold for your own skill implementation.
 license: Apache-2.0
 aea_version: '>=1.0.0, <2.0.0'
 fingerprint:
-  __init__.py: bafybeicg2egktf6eucqvo76fekse52m6iofdrdgq2636imky4ccgmqeatm
-  behaviours.py: bafybeieyy5pq534tyxgjyabfgbv3xbqxnuyt4welc4hgns37ldqa73z6de
-  dialogues.py: bafybeicvlr5dctljsa76itiyop3v7hc6frymuanitmlnxhs5cnnwyey3fu
-  handlers.py: bafybeiduxt7jqbnjcz2spxppyx4qzranow44a624nibtprd43x4ev3355u
-  models.py: bafybeibgthksq6qitsukfby7b6xa7nhaubtmtv6hdfzfsvei6uozrm7bgm
-  orchestrator_fsm.yaml: bafybeihdeh4s7ohlfhwdo3xxkod2xvkkr3lddpky5cg7btknjaqgyx6ra4
+  __init__.py: bafybeie2eua5twi3crfg7cqxkv62glygjnyaxlncasxsqqep3xibhsspiq
+  behaviours.py: bafybeibu73du4bptw45fsf7ddpel7ddkguobzbbfh4uxqzvpu2ogn7n6he
+  dialogues.py: bafybeihyy54kdw66khuvielotnhskgfepllwbwh4df4eqonprywbnivfku
+  fsm_specification.yaml: bafybeihdeh4s7ohlfhwdo3xxkod2xvkkr3lddpky5cg7btknjaqgyx6ra4
+  handlers.py: bafybeicwypeetjetfnkdtz5k6rq7jifa65zkv5yu5rs4mrsigqofiuizty
+  models.py: bafybeigpcx34wrocfo7yeizjoqmj4abvypcytmfiv5h3rzrkac4cwe5wsu
   payloads.py: bafybeic4xjaksu3bxuiduqhispjiculoyfrgw2bf7c4tsmcyuksseehr2e
-  rounds.py: bafybeic2dmak3763adpgwz4ajgbzdi3u32hbikgzjdmdntsjzist6gk5tm
+  rounds.py: bafybeiha4bm4tepfayt5ropnaiuxgypskkmxxkxdj2eprvueukchvgayg4
   tests/__init__.py: bafybeifigp64li3j3yidpan5arc27etm3jytadjsckjaidrem57hst73ry
-  tests/test_behaviours.py: bafybeicud5ffismvxcyhprdb4ml4xthw2bn5hhp4zudggaluvnstjebcsy
-  tests/test_dialogues.py: bafybeih7kqynur3k26hoegkpqn4cuunjo5kc3ar4mf76fz7ipvzlqf6cdm
-  tests/test_handlers.py: bafybeiexy3kzzyyojjmpqrnpvcwhtycynzaudzgvuxzr4lehp3ppwvufsy
-  tests/test_models.py: bafybeihe6wlknwa5pzdkxuat77te7mnsjhyafu3wws2geekdfx3lpc7f3q
-  tests/test_payloads.py: bafybeiccin6flmgnqh5guvurfkcvgd2mnj76rz7zabysbcu2or5h4ase6i
-  tests/test_rounds.py: bafybeietd4ezu2ge4em4auszhlumrklzkujesvs72xpt56bgrbjctxwazq
+  tests/test_behaviours.py: bafybeignr3x3coilwgbk2i4xbo3al7juu3pqwmzumprddv6owmzy2rdvzi
+  tests/test_dialogues.py: bafybeiehikfibq3gbxkkwfbbzlqykr7wvlvcrar3t6tnddwzexz6hgg3h4
+  tests/test_handlers.py: bafybeie3kjcyj4m3yg4a25eiivuutvidwco4qj4nzkgxyo6tgrv2slau4u
+  tests/test_models.py: bafybeibviskggtgagrgvq6k4esy6hchwxidy3x7wmcstughcemhdslpvb4
+  tests/test_payloads.py: bafybeid2ml5xuwlqbvizgvags5vt7hnbpdurbl7doswqk3ka3egoqptx5i
+  tests/test_rounds.py: bafybeih32ojpdfrg7qqw4suqe7h3gqx747ckjxnndcpl5vm4nxhxpxt2ji
 fingerprint_ignore_patterns: []
 connections: []
 contracts: []
 protocols: []
 skills:
-- valory/abstract_round_abci:0.1.0
+- valory/abstract_round_abci:0.1.0:bafybeibovsktd3uxur45nrcomq5shcn46cgxd5idmhxbmjhg32c5abyqim
 behaviours:
   main:
     args: {}
-    class_name: OrchestratorRoundBehaviour
+    class_name: QSOrchestratorRoundBehaviour
 handlers:
   abci:
     args: {}
@@ -100,6 +100,7 @@ models:
       ipfs_domain_name: null
       keeper_allowed_retries: 3
       keeper_timeout: 30.0
+      light_slash_unit_amount: 5000000000000000
       max_attempts: 10
       max_healthcheck: 120
       on_chain_service_id: null
@@ -110,7 +111,8 @@ models:
       retry_attempts: 400
       retry_timeout: 3
       round_timeout_seconds: 30.0
-      service_id: orchestrator
+      serious_slash_unit_amount: 8000000000000000
+      service_id: qs_solver
       service_registry_address: null
       setup:
         all_participants:
@@ -118,6 +120,8 @@ models:
         consensus_threshold: null
         safe_contract_address: '0x0000000000000000000000000000000000000000'
       share_tm_config_on_startup: false
+      slash_cooldown_hours: 3
+      slash_threshold_amount: 10000000000000000
       sleep_time: 1
       tendermint_check_sleep_delay: 3
       tendermint_com_url: http://localhost:8080
@@ -125,6 +129,8 @@ models:
       tendermint_p2p_url: localhost:26656
       tendermint_url: http://localhost:26657
       tx_timeout: 10.0
+      use_slashing: false
+      use_termination: false
       validate_timeout: 1205
     class_name: Params
   requests:
@@ -140,5 +146,5 @@ models:
     args: {}
     class_name: TendermintDialogues
 dependencies: {}
-is_abstract: false
+is_abstract: true
 customs: []
